Allow 'click' to be given a user param
* Create utils module to contain update_environ function
* Override response click method to call update_environ
